+++ vips/vips.h
@@ -19,7 +19,7 @@ int vips_pngload_go(void *buf, size_t len, VipsImage **out);
 int vips_webpload_go(void *buf, size_t len, double scale, int pages, VipsImage **out);
 int vips_gifload_go(void *buf, size_t len, int pages, VipsImage **out);
 int vips_svgload_go(void *buf, size_t len, double scale, VipsImage **out);
-int vips_heifload_go(void *buf, size_t len, VipsImage **out);
+int vips_heifload_go(void *buf, size_t len, VipsImage **out, int thumbnail);
 int vips_tiffload_go(void *buf, size_t len, VipsImage **out);
 
 int vips_black_go(VipsImage **out, int width, int height, int bands);