+++ processing_handler.go
@@ -112,6 +112,8 @@ func respondWithImage(reqID string, r *http.Request, rw http.ResponseWriter, sta
 		rw.Header().Set("X-Origin-Content-Length", strconv.Itoa(len(originData.Data)))
 		rw.Header().Set("X-Origin-Width", resultData.Headers["X-Origin-Width"])
 		rw.Header().Set("X-Origin-Height", resultData.Headers["X-Origin-Height"])
+		rw.Header().Set("X-Result-Width", resultData.Headers["X-Result-Width"])
+		rw.Header().Set("X-Result-Height", resultData.Headers["X-Result-Height"])
 	}
 
 	rw.Header().Set("Content-Length", strconv.Itoa(len(resultData.Data)))