+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Add
 - Add support of RLE-encoded BMP.
 - Add `IMGPROXY_ENFORCE_THUMBNAIL` config and [enforce_thumbnail](https://docs.imgproxy.net/generating_the_url?id=enforce-thumbnail) processing option.
+- Add `X-Result-Width` and `X-Result-Height` to debug headers.
 
 ### Change
 - Use thumbnail embedded to HEIC/AVIF if its size is larger than or equal to the requested.